refactor: 코리 반영(모달 제거, index폴더 교체, style수정)
