+++ src/app/signup/page.tsx
@@ -1,28 +1,23 @@
 "use client";
 
-import Modal from "@/common/component/Modal";
-import AuthHeader from "@/view/index/AuthHeader";
-import Footer from "@/view/index/Footer";
+import AuthHeader from "@/shared/component/AuthHeader";
+import FormFooter from "@/shared/component/FormFooter";
+import { wrapper } from "@/view/login/index.css";
 import SignupForm from "@/view/signup/SignupForm";
 import { containerStyle } from "@/view/signup/index.css";
-import { useRouter } from "next/navigation";
 
 const SignupPage = () => {
-  const router = useRouter();
-  const onClose = () => {
-    router.push("/");
-  };
   return (
-    <Modal isOpen={true} onClose={onClose}>
-      <AuthHeader handleClick={onClose} />
+    <div className={wrapper}>
+      <AuthHeader />
       <div className={containerStyle}>
         <SignupForm />
-        <Footer
+        <FormFooter
           guideLabel="이미 계정이 있으신가요?"
           link={{ href: "/login", label: "로그인하기" }}
         />
       </div>
-    </Modal>
+    </div>
   );
 };
