+++ src/shared/component/AuthHeader/index.css.ts
@@ -4,15 +4,12 @@ import { recipe } from "@vanilla-extract/recipes";
 
 export const headerStyle = recipe({
   base: {
-    position: "relative",
-    zIndex: 0,
     display: "flex",
     justifyContent: "space-between",
     alignItems: "center",
 
     width: "100vw",
     height: "7.2rem",
-    margin: 0,
     padding: "0 3.2rem 0 6.4rem",
 
     backgroundColor: theme.color.bg,
