+++ src/shared/component/Card/index.css.ts
